+++ src/main/java/io/blamer/bot/conversation/package-info.java
@@ -22,37 +22,9 @@
  * SOFTWARE.
  */
 
-package io.blamer.bot.bot.initializer;
-
-import io.blamer.bot.bot.Bot;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.event.ContextRefreshedEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
 /**
- * Bot initializer.
+ * @author Aliaksei Bialiauski (abialiauski.dev@gmail.com)
+ * @since 0.0.0
  */
-@Component
-@RequiredArgsConstructor
-public class BotInitializer {
 
-  /**
-   * Bot.
-   */
-  private final Bot bot;
-
-  /**
-   * Initializes the bot.
-   *
-   * @throws TelegramApiException if fails.
-   */
-  @EventListener({ContextRefreshedEvent.class})
-  public void initializeBot() throws TelegramApiException {
-    final TelegramBotsApi api = new TelegramBotsApi(DefaultBotSession.class);
-    api.registerBot(this.bot);
-  }
-}
+package io.blamer.bot.conversation;
