--- src/test/java/io/blamer/bot/BotApplicationEntryTest.java
@@ -1,13 +1,0 @@
-package io.blamer.bot;
-
-import annotation.TestWithSpringContext;
-import org.junit.jupiter.api.Test;
-
-@TestWithSpringContext
-class BotApplicationEntryTest {
-
-	@Test
-	void loadsContext() {
-	}
-
-}
